feat: improve rust bindings compat (#395)
improve rust bindings compat, enabling syntaxes like
- segments.memory.get(idx)
or 
- dict_trackers[idx] = X

based on a custom fork of the cairo vm in which I unbotherly created a
special `DictKey` type which is an
enum

```rust
#[derive(
    Eq, Ord, Hash, PartialEq, PartialOrd, Clone, Debug,
)]
pub enum DictKey {
    Simple(MaybeRelocatable),
    Compound(Vec<MaybeRelocatable>),
}
```

Which allows a Vec of Felts / Relocatable to be used as dict keys. For
us this is important because we need to track preimages that might not
fit in a felt inside the dict tracker.

This was implemented in a "dumb" way that is unconsidering of external
side effects to the Cairo-VM - once synced with LC we will want to
revisit the approach. see
https://github.com/kkrt-labs/cairo-vm/tree/kkrt-compat