dev: add cursor directory (#864)
trying stuff with claude 3.7 x Claude. This should give the composer
mode a better grasp of our codebase, patterns, usages, and be able to do
better things in agentic mode. I hope we'll be able to speedup the
refactoring to proper typings with LLMs.

See https://www.cursor.com/changelog