dev: expose `serialize` in RustVM pythonic hints (#887)
Exposes a `serialize` function, available in the scope of all hints,
that enables the serialization using `serde` of cairo types, and returns
their python representation.

This can be used for the logging / tracing system.

I am welcoming any alternative idea on the currently proposed design -
as working with python `globals()` has been quite a hassle.

Also, i'm not sure yet how we can make this entirely transparent with
the pythonVM runs. If we realize there's a need for that, we can do it
later.

This does introduce a small performance overhead of about ~5% from what
i've observed empirically. So in production we'll need to make sure
we're not compiling with this flag.
Closes #883

---------

Co-authored-by: Clément Walter <clement0walter@gmail.com>