Improve simplify_jumpdest_check_push_last_32_bytes rust hint
The hint takes ~30% of time execution when proving a block.

Currently, the function calls serialize_sequence on bytecode,
which is inefficient since we only need 32 bytes at max.

Instead, we craft a pointer to the start of the relevant memory address
(until jump dest key value), and build the u8 vector necessary to check
is_no_push