fix: handle deleted accounts in a sound way (#1378)
Closes https://github.com/kkrt-labs/keth/issues/1108
Closes https://github.com/kkrt-labs/keth/issues/1382 (necessary fix in
this pr)

Several things addressed in this PR. The main goal is to fix the way we
handle deleted account, which currently is unsound. I started by
identifying all such cases where we might encounter these cases of
deletions. One thing to be careful about are rules remnant from outdated
EVM forks, and the blocks >= Cancun we're going to prove.

**Note:** It has been confirmed that, Post-Merge, there exists no
account in the ethereum state that is empty but has storage.

### Remove storage destructions on empty accounts

As per [EIP-161](https://eips.ethereum.org/EIPS/eip-161)

>  Notes
> In the present Ethereum protocol, it should be noted that very few
state changes can ultimately result in accounts that are empty following
the execution of the transaction. In fact there are only four contexts
that current implementations need track:
> 
> an empty account has zero value transferred to it through CALL;
> an empty account has zero value transferred to it through SUICIDE;
> an empty account has zero value transferred to it through a
message-call transaction;
> an empty account has zero value transferred to it through a
zero-gas-price fees transfer.

Related to: calling `destroy_account` on `coinbase` and
`touched_empty_accounts`:
a. `get_account_optional` can return None or an Account
b. `trie_set` called with an EMPTY_ACCOUNT will delete this account from
the trie
-> Thus, `get_account_optional` should theoritically never return an
EMPTY_ACCOUNT, but always either None or an existing account.

**EXCEPT IF**: the account was created before
[EIP-156](https://eips.ethereum.org/EIPS/eip-158) in which case it can
be empty and have some storage, in which case, the storage is cleared.
Because we have access to the storage roots, we can include them in our
check for EMPTY_ACCOUNT, which will make sure that, if the account is
empty but has a non-empty storage root, we must destroy its storage.

It would be interesting to know whether some of these accounts still
exist or not. Basically, this is a "self-cleaning" mechanism for the
ethereum protocol to remove these old accounts.


### Handling storage deletions for SELFDESTRUCT-same-tx accounts

Accounts that are created in a TX and destructed in the same TX do not
produce any state diff - and storage ends up being cleared. This means
that:
- Any eventually touched storage key is in `accessed_storage_keys` of
that transaction
- Instead of looking for the storage keys associated to this account in
the State, we can look for them in `accessed_storage_keys`.
- This will reduce the amount of iterations performed.

### Handling of CREATE on contracts that already have storage (remnants
of EIP-161, EIP-158)

As we saw previously, there exists accounts that are empty, BUT have
state. As such, the behavior of `process_create_message` is to, by
default, erase any existing state associated to an account.
An empty account can only have storage if it's a remnant of an old fork,
meaning that it had not been created in the current block. As such, we
have access to its storage root as per ZKPI inputs. Meaning that:
- If that storage root is the empty root, we can simply do nothing
- If that storage root is not empty, we must delete the existing
storage.


From all these conditions, we can identify two distinct conditions:

- Storage destruction on empty accounts with storage that were created
**before** the current block
- Storage destructions on accounts created in the same TX, and then
selfdestructed.

For 1. 
- If the account has a storage root, then, we need to iterate on the
state to delete any associated storage key. We're ensured these keys are
present in the state, because ZKPI provides them (as they're _all_
deleted)
- If the account doesn't have a storage root, then, we have nothing to
do, besides setting the account to `None`.

For 2.
- We can simply iterate of `accessed_storage_keys` and set the storage
value to 0.